+++ topothehourbot/utilities/ircv3/channeling/__init__.py
@@ -1,3 +1,3 @@
-from .core import *
+from .channels import *
 from .protocols import *
 from .series import *
